+++ restapi/embedded_spec.go
@@ -56,7 +56,7 @@ func init() {
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "version": "2.3.1",
+    "version": "2.4.0",
     "x-logo": {
       "altText": "Archer logo",
       "backgroundColor": "#FFFFFF",
@@ -2398,7 +2398,7 @@ func init() {
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "version": "2.3.1",
+    "version": "2.4.0",
     "x-logo": {
       "altText": "Archer logo",
       "backgroundColor": "#FFFFFF",